[#ICC-188] rename mock durable-functions.ts to context.ts
