+++ __mocks__/context.ts
@@ -1,0 +1,8 @@
+export const context = {
+  log: {
+    error: jest.fn(),
+    info: jest.fn(),
+    verbose: jest.fn(),
+    warn: jest.fn()
+  }
+};
